+++ DisCatSharp.ApplicationCommands/Attributes/ContextMenu/ContextMenuCooldownAttribute.cs
@@ -22,11 +22,10 @@
 
 using System;
 using System.Collections.Concurrent;
-using System.Globalization;
-using System.Threading;
 using System.Threading.Tasks;
 
 using DisCatSharp.ApplicationCommands.Context;
+using DisCatSharp.ApplicationCommands.Entities;
 using DisCatSharp.ApplicationCommands.Enums;
 
 namespace DisCatSharp.ApplicationCommands.Attributes;
